REFACTOR(DTO): Rename User DTO's
- Renamed `UserUpdateRequest` to `UserUpdateDTO` for consistency.

-Renamed `UserCreateRequest` to `UserCreateDTO` for consistency.

- Moved `UserResponseDTO` to **dto.user** package for better code organization.
